+++ CHANGELOG.md
@@ -21,3 +21,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Setup zod
 * Setup typst
 * Install xlxs (SheetJS)
+* Setup nodemailer